Migrate fjs/sul to .f.mjs, split types into types.ts
Splits the EncodeState<S>/Encode<S> types into a sibling types.ts, and
converts the runtime implementation (emptyEncodeState, encode) to
JSDoc-typed module.f.mjs/proof.f.mjs under the stage-1 TypeScript-to-mjs
migration. No other module imports fjs/sul/module.f.ts, so updates only
touch the README.md example.
