+++ CHANGELOG.md
@@ -20,6 +20,12 @@ history.
 
 ## Unreleased
 
+- **BREAKING CHANGES:** `fjs/sul` migrates from authored TypeScript
+  (`.f.ts`) to JSDoc-typed JavaScript (`.f.mjs`), splitting the
+  `EncodeState`/`Encode` types into a sibling `types.ts` — importers
+  must use the `.f.mjs` specifier for runtime values and the `types.ts`
+  specifier for types
+  [#TBD](https://github.com/functionalscript/functionalscript/pull/TBD)
 - **BREAKING CHANGES:** `fjs/types/object`, `fjs/sul/level/literal`,
   `fjs/sul/id`, `fjs/sul/level/hash`, `fjs/types/string_set`, and the
   `module.f.ts` of `fjs/types/sorted_set` and `fjs/types/byte_set`
